backend/requirements/base.txt
Django==4.2.16
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
drf-spectacular==0.27.2
django-filter==24.3
psycopg2-binary==2.9.9
Pillow==10.4.0
python-decouple==3.8
celery==5.4.0
redis==5.0.8
django-redis==5.4.0
cloudinary==1.41.0
django-cloudinary-storage==0.3.0
django-ratelimit==4.1.0
sentry-sdk==2.14.0
whitenoise==6.7.0
gunicorn==22.0.0
